requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.0.0
channels-redis==4.1.0
chardet==5.2.0
charset-normalizer==3.4.1
cloudinary==1.43.0
constantly==23.10.4
cryptography==44.0.2
daphne==4.0.0
dj-database-url==2.3.0
Django==5.1.2
django-cloudinary-storage==0.3.0
django-paypal==2.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
gunicorn==23.0.0
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonfield==3.1.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
msgpack==1.1.0
packaging==24.2
pillow==10.4.0
priority==1.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
PyPDF2==3.0.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
referencing==0.36.2
reportlab==4.2.5
requests==2.32.3
rpds-py==0.25.1
service-identity==24.2.0
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.8.2
zope.interface==7.2
